refactor(voice/health): split cascade.py into subpackage [AP-16, T1.2]
Closes anti-pattern #16 hygiene for cascade.py (1925 LOC, mixed
responsibilities — 4 platform builders + 4 public tuples + 4 lookups
+ 1 ProbeCallable Protocol + executor with 600 LOC core loop +
candidate-set runner + 5 logging helpers + 2 helpers + budget/
quarantine/lock-singleton).

Layout per godfile-splits mission Part 3 / T02:

  voice/health/cascade/
    __init__.py     — re-exports the 8 public symbols + 2 internal
                      helpers used by tests (`_platform_cascade`,
                      `_run_mixer_sanity` — both noqa F401)
    _planner.py     — 4 cascade builders + 4 public tuples
                      (WINDOWS_CASCADE, WINDOWS_CASCADE_AGGRESSIVE,
                      LINUX_CASCADE, MACOS_CASCADE) +
                      build_linux_cascade_for_device + _platform_cascade
                      + _PLATFORM_CASCADES
    _alignment.py   — _lookup_override + _lookup_store + _run_mixer_sanity
                      (pre-cascade lookups + L2.5 invocation)
    _budget.py      — 5 tuning constants (sourced from VoiceTuningConfig
                      per AP-17) + _DEFAULT_LOCKS singleton + _default_locks
                      + _quarantine_endpoint + _record_winner
    _executor.py    — ProbeCallable Protocol + _call_probe + run_cascade
                      + _run_cascade_locked (the 474-LOC core loop) +
                      _try_combo + run_cascade_for_candidates + 5 logging
                      helpers + _LOG_DETAIL_MAX_CHARS

Cross-submodule deps are forward-only — no cycles:

  _planner    → contract.Combo + VoiceTuningConfig
  _alignment  → contract types + (lazy) _mixer_sanity
  _budget     → engine._lock_dict + contract types + _metrics
  _executor   → _planner + _alignment + _budget + contract + probe + _quarantine

Test patch migration (anti-pattern #20):

  Wave 3 recon reported 0 patches against `cascade.X` — DRIFTED. At
  HEAD a fresh grep found 12 patch sites in
  `tests/unit/voice/health/test_cascade_mixer_sanity.py` targeting
  `sovyx.voice.health.cascade._run_mixer_sanity`. Migrated all 12 to
  `sovyx.voice.health.cascade._executor._run_mixer_sanity` (the
  binding the executor reads — patching `_alignment._run_mixer_sanity`
  would NOT take effect because executor's `from ._alignment import`
  already snapshots the reference at import time, exactly as
  documented in CLAUDE.md AP-20).

  Same test also calls `cascade_mod._run_mixer_sanity(...)` directly
  (3 sites). Re-exporting `_run_mixer_sanity` in the package
  `__init__.py` (with `# noqa: F401`) keeps those direct calls working
  unchanged — this is intentional API preservation, not a public
  commitment.

Spec deviation: Wave 3 recon's "0 test patches on cascade" claim was
stale. 12 patches found at HEAD; all migrated in this commit per AP-20.

Validation:
  * 8-name smoke import: OK
  * ruff check + format clean
  * mypy strict: 412 source files, 0 issues
  * bandit clean
  * pytest full suite: 11768 passed, 25 skipped, 57 deselected
    (identical to pre-split baseline)

Refs: docs-internal/missions/MISSION-voice-godfile-splits-v0.24.1.md
      Part 3 / T02
      docs-internal/missions/MISSION-voice-final-skype-grade-2026.md
      Phase 1 / T1.2
      CLAUDE.md anti-pattern #16, #20
